wayland: Implement xx-pip-v1
The xx-pip-v1 protocol provides clients a way to create floating windows
with miniature contents, for example a video (a movie or a video call),
a map with directions, a timer countdown, etc.

These windows are placed in the overlay layer above all windows, including
fullscreen windows.

This is an experimental version of the protocol.