ButtonRebindsFilter: Add support for rebinding to tablet dials
This builds on top of the existing tablet dial support, and allows users
to configure how we rebind them. Since dial support is so new in Linux,
it's not expected for every application to support them right away.

To keep it simple, the only supported rebind target (apart from being
disabling) is to the vertical scroll wheel. These dials are commonly
used to zoom or change brush sizes, for example which can be done
through emulated scroll wheel events.