scripting: Fix dbus calls with numeric types
DBus calls with numeric types other than int and
double were not working(eg. uint16, int16, etc).

This was because the JavaScript types were
converted by QJSValue::toVariant(), which only
supports int and double.

This change adds support for the other numeric
types which JavaScript does not support by
introspecting the type signature of the DBus
method and converting the JavaScript value to
the appropriate QVariant type.

This is important because if one tries to call a
DBus method with an incorrect type signature, the
call will fail.

BUG: 486024