Don't migrate fullscreen windows between layers
The NETWM spec recommends to put active fullscreen windows above windows
with the above state and docks. This is mainly to hide panels.

On the other hand, we received a bunch of complaints throughout years
about windows with the keep above state being below fullscreen windows.

The slide effect has its own workarounds to force dock windows on top of
fullscreen windows too.

It also prevents the dock functioning as expected when there's a
fullscreen window.