compositor: implement software cursors differently
Instead of having a separate layer with its own framebuffer, we add a CursorItem
to the scene. While a hardware cursor is used, we then hide that item from the
scene on a per-output / delegate basis, which both allows us to delete a bunch
of cursor specific code, and prepares for similar overlays in the future.